refactor: move server to root
